Refactor name of theme provider
