+++ packages/pwa/pages/_app.page.tsx
@@ -6,7 +6,7 @@ import { appWithTranslation } from 'next-i18next'
 import { GlobalStyles } from 'twin.macro'
 import { Toast } from '@forest-restoration/shared'
 import { useFixHeightViewport } from '../shared/hooks/useFixHeightViewport'
-import { ThemeProvider, useThemeProvider } from '../shared/providers/themeProvider'
+import { ThemeProvider, useThemeProvider } from '../shared/providers/theme'
 import { Navigation } from '../shared/components/navigation'
 import '../theme.css'
 import '../firebase/clientApp'
